update debian init script to be more robost
This adds configurability via the standard /etc/default/newrelic_plugin_agent files.
Add --chuid to make the agent run as the newrelic user by default. This is configurable with the RUNAS parameter.
Add --retry when stopping the agent so that after TIMEOUT (default=5) the process is killed. Previously there was no check at all on wether or not the agent actually exited.